requirements.txt
streamlit 
pandas
numpy
matplotlib.pyplot
seaborn